Shifted a query
Have shifted a query used to populate variable "v_mview_table_alias". Ideally the variable value should not be changed/garbled by the functions which are being called post to the variable initialization, but here I was facing such issues and hence moved the query to immediate upper line where the variable is actually being used.